# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prasso Invenbin API",
        "description": "API endpoints for Prasso Invenbin",
        "contact": {
            "email": "bcp@faxt.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Retrieve a list of product records with eager loading of products relationship",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "guid": {
                                                "type": "string",
                                                "example": "11111-11111-11111-11111"
                                            },
                                            "product_name": {
                                                "type": "string",
                                                "example": "Product A"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "01/01/2024"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product",
                "operationId": "eddf8c7847239ddc1040daadd814e21b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/products/{guid}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Display the specified resource.",
                "operationId": "6d124b5635ad34a7d53425151b0b3783",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductInputOutput"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update a product",
                "operationId": "5bda4aa29bbba8a879a0bc741f37b8a6",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Remove the specified resource from storage.",
                "operationId": "37b5ae0cc0c522aa69f759ba1c706c88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductInputOutput": {
                "title": "ErpProduct",
                "description": "Erp Product model",
                "properties": {
                    "sku": {
                        "description": "SKU of the product",
                        "type": "string"
                    },
                    "guid": {
                        "description": "GUID of the product",
                        "type": "string"
                    },
                    "product_name": {
                        "description": "Name of the product",
                        "type": "string"
                    },
                    "short_description": {
                        "description": "Short description of the product",
                        "type": "string"
                    },
                    "attribute_xml": {
                        "description": "XML attributes of the product",
                        "type": "string"
                    },
                    "stock_location": {
                        "description": "Location of the product in stock",
                        "type": "string"
                    },
                    "our_price": {
                        "description": "Price of the product set by us",
                        "type": "number",
                        "format": "decimal"
                    },
                    "retail_price": {
                        "description": "Retail price of the product",
                        "type": "number",
                        "format": "decimal"
                    },
                    "weight": {
                        "description": "Weight of the product",
                        "type": "number",
                        "format": "decimal"
                    },
                    "currency_code": {
                        "description": "Currency code of the product price",
                        "type": "string"
                    },
                    "unit_of_measure_id": {
                        "description": "IDUnit of measure for the product",
                        "type": "string"
                    },
                    "admin_comments": {
                        "description": "Comments from the admin about the product",
                        "type": "string"
                    },
                    "length": {
                        "description": "Length of the product",
                        "type": "number",
                        "format": "decimal"
                    },
                    "height": {
                        "description": "Height of the product",
                        "type": "number",
                        "format": "decimal"
                    },
                    "width": {
                        "description": "Width of the product",
                        "type": "number",
                        "format": "decimal"
                    },
                    "dimension_unit_id": {
                        "description": "ID Unit of measure for dimensions of the product",
                        "type": "string"
                    },
                    "list_order": {
                        "description": "Order of the product in the list",
                        "type": "integer"
                    },
                    "rating_sum": {
                        "description": "Sum of ratings for the product",
                        "type": "integer"
                    },
                    "total_rating_votes": {
                        "description": "Total number of rating votes for the product",
                        "type": "integer"
                    },
                    "default_image": {
                        "description": "URL or path to the default image of the product",
                        "type": "string"
                    },
                    "owned_by": {
                        "description": "ID of the owner of the product",
                        "type": "integer"
                    },
                    "inventory_count": {
                        "description": "Inventory count of the product",
                        "type": "integer"
                    },
                    "reorder_point": {
                        "description": "Reorder point of the product",
                        "type": "integer"
                    },
                    "product_status_id": {
                        "description": "ID of the status of the product",
                        "type": "integer"
                    },
                    "product_type_id": {
                        "description": "ID of the type of the product",
                        "type": "integer"
                    },
                    "updated_by": {
                        "description": "User who last updated the product",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}